Removed unused Grunt helpers. Removed unused deprecated modules from versioned meta in desktop application
