+++ lib/helpers.js
@@ -1,95 +1,27 @@
 'use strict';
 
 const minimatch = require('minimatch');
-const async = require('async');
 const path = require('path');
-const fs = require('fs-extra');
 const zlib = require('zlib');
 const logger = require('./logger').logger();
 
 const isWindows = process.platform === 'win32';
-
-const writeOptions = { flag: 'wx' };
 const gzOptions = {
    level: zlib.Z_BEST_COMPRESSION,
    strategy: zlib.Z_DEFAULT_STRATEGY
 };
 const dblSlashes = /\\/g;
-
-function copyFile(target, dest, dataToCopy, cb) {
-   if (dataToCopy) {
-      writeFile(dest, dataToCopy, cb);
-   } else {
-      fs.readFile(target, (err, data) => {
-         if (err) {
-            logger.error({
-               error: err
-            });
-            cb();
-         } else {
-            writeFile(dest, data, cb);
-         }
-      });
-   }
-}
-
-function mkSymlink(target, dest, cb) {
-   const link = function(linkTarget, linkDest, linkCb) {
-      fs.symlink(linkTarget, linkDest, (linkErr) => {
-         if (linkErr && linkErr.code === 'ENOENT') {
-            fs.ensureDir(path.dirname(linkDest), (err) => {
-               if (!err || err.code === 'EEXIST') {
-                  link(linkTarget, linkDest, linkCb);
-               } else {
-                  logger.error({
-                     error: err
-                  });
-                  linkCb();
-               }
-            });
-         } else if (linkErr && linkErr.code !== 'EEXIST') {
-            logger.error({
-               error: linkErr
-            });
-            linkCb();
-         } else {
-            linkCb();
-         }
-      });
-   };
-
-   link(target, dest, cb);
-}
-
-function recurse(input, handler, done, limit) {
-   fs.readdir(input, (err, files) => {
-      if (!err) {
-         async.eachLimit(
-            files,
-            limit || 20,
-            (file, cb) => {
-               const prettyFilePath = file.normalize();
-               const abspath = path.join(input, prettyFilePath);
-
-               fs.lstat(abspath, (statErr, stats) => {
-                  if (!statErr) {
-                     if (stats.isDirectory()) {
-                        recurse(abspath, handler, cb, Math.floor(limit / 1.2) || 1);
-                     } else {
-                        handler(abspath, cb);
-                     }
-                  } else {
-                     cb(statErr);
-                  }
-               });
-            },
-            (limitErr) => {
-               done(limitErr);
-            }
-         );
-      }
-   });
-}
+const deprecatedControlsToSave = [
+   'ModalOverlay',
+   'LoadingIndicator',
+   'ProgressBar',
+   'FieldString',
+   'LinkButton',
+   'Button',
+   'FieldAbstract',
+   'ButtonAbstract',
+   'Menu'
+];
 
 function validateFile(file, patterns) {
    let passed = false;
@@ -111,10 +43,6 @@ function validateFile(file, patterns) {
    return passed;
 }
 
-function percentage(x, all) {
-   return Math.floor(x * 100 / all);
-}
-
 function sortObject(obj, comparator) {
    const sorted = {};
    Object.keys(obj)
@@ -130,43 +58,6 @@ function sortObject(obj, comparator) {
          }
       });
    return sorted;
-}
-
-function writeFile(dest, data, cb) {
-   fs.writeFile(dest, data, writeOptions, (writeErr) => {
-      if (writeErr && writeErr.code === 'ENOENT') {
-         fs.ensureDir(path.dirname(dest), (err) => {
-            if (!err || err.code === 'EEXIST') {
-               writeFile(dest, data, cb);
-            } else {
-               logger.error({
-                  error: err
-               });
-               cb();
-            }
-         });
-      } else if (writeErr && writeErr.code !== 'EEXIST') {
-         logger.error({
-            error: writeErr
-         });
-         cb();
-      } else {
-         cb();
-      }
-   });
-}
-
-function writeGzip(dest, data, callback) {
-   zlib.gzip(data, gzOptions, (err, compressed) => {
-      if (!err) {
-         writeFile(dest, compressed, callback);
-      } else {
-         logger.error({
-            error: err
-         });
-         callback();
-      }
-   });
 }
 
 function gzip(data) {
@@ -249,19 +140,6 @@ function removeLatestSlash(filePath) {
    return filePath;
 }
 
-async function tryRemoveFolder(folder) {
-   try {
-      await fs.remove(folder);
-   } catch (e) {
-      return e;
-   }
-   return null;
-}
-
-const delay = function(ms) {
-   return new Promise(resolve => setTimeout(resolve, ms));
-};
-
 const promisifyDeferred = function(deferred) {
    return new Promise((resolve, reject) => {
       deferred
@@ -320,24 +198,47 @@ function removeFileFromVersionedMeta(versionedMeta, fullPath) {
    }
 }
 
+/**
+ * Проверяем WS.Deprecated модули, нам необходимо оставить в папке
+ * Controls только неймспейсы deprecatedControlsToSave
+ * @param{String} fullPath - полный путь до копируемого файла
+ * @returns {boolean} true - не удаляем. false - удаляем.
+ */
+function checkPathForDeprecatedToRemove(fullPath) {
+   if (fullPath.includes('WS.Deprecated/Controls')) {
+      let result = false;
+      deprecatedControlsToSave.forEach((currentControl) => {
+         if (fullPath.includes(`WS.Deprecated/Controls/${currentControl}/`)) {
+            result = true;
+         }
+      });
+      return result;
+   }
+   return true;
+}
+
+/**
+ * Проверяем, нужно ли удалить модуль из оффлайн-приложения.
+ * @param{String} prettyPath - путь до файла
+ * @returns {boolean}
+ */
+function needToRemoveModuleForDesktop(prettyPath) {
+   return !checkPathForDeprecatedToRemove(prettyPath) ||
+      prettyPath.includes('SBIS3.CONTROLS/themes/online') ||
+      prettyPath.includes('WS.Core/lib/Ext');
+}
+
 module.exports = {
-   writeFile,
-   writeGzip,
-   copyFile,
-   mkSymlink,
-   recurse,
    validateFile,
-   percentage,
    sortObject,
    getFirstDirInRelativePath,
    prettifyPath,
    unixifyPath,
    removeLeadingSlash,
    removeLatestSlash,
-   tryRemoveFolder,
-   delay,
    promisifyDeferred,
    gzip,
    isEqualObjectFirstLevel,
-   removeFileFromVersionedMeta
+   removeFileFromVersionedMeta,
+   needToRemoveModuleForDesktop
 };
